+++ atlasdb-client/src/test/java/com/palantir/atlasdb/table/description/SchemaTest.java
@@ -15,6 +15,7 @@
  */
 package com.palantir.atlasdb.table.description;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.not;
@@ -23,12 +24,17 @@ import static org.junit.Assert.assertThat;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
+import org.apache.commons.lang.StringUtils;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 import com.google.common.io.Files;
+import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.keyvalue.api.Namespace;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 
@@ -76,6 +82,57 @@ public class SchemaTest {
                         containsString("import java.util.Optional")));
     }
 
+    @Test
+    public void testIgnoreTableNameLengthFlag() throws IOException {
+        Schema schema = new Schema("Table", TEST_PACKAGE, Namespace.EMPTY_NAMESPACE);
+        schema.ignoreTableNameLengthChecks();
+        String longTableName = String.join("", Collections.nCopies(1000, "x"));
+        TableReference tableRef = TableReference.createWithEmptyNamespace(longTableName);
+        schema.addTableDefinition(longTableName, getSimpleTableDefinition(tableRef));
+    }
+
+    @Test
+    public void testLongTableNameLengthFailsCassandra() throws IOException {
+        Schema schema = new Schema("Table", TEST_PACKAGE, Namespace.EMPTY_NAMESPACE);
+        int longLengthCassandra = AtlasDbConstants.CASSANDRA_TABLE_NAME_CHAR_LIMIT + 1;
+        String longTableName = String.join("", Collections.nCopies(longLengthCassandra, "x"));
+        TableReference tableRef = TableReference.createWithEmptyNamespace(longTableName);
+        List<CharacterLimitType> kvsList = new ArrayList<CharacterLimitType>();
+        kvsList.add(CharacterLimitType.CASSANDRA);
+        assertThatThrownBy(() ->
+                schema.addTableDefinition(longTableName, getSimpleTableDefinition(tableRef)))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(getErrorMessage(longTableName, kvsList));
+    }
+
+    @Test
+    public void testLongTableNameLengthFailsBoth() throws IOException {
+        Schema schema = new Schema("Table", TEST_PACKAGE, Namespace.EMPTY_NAMESPACE);
+        String longTableName = String.join("", Collections.nCopies(1000, "x"));
+        TableReference tableRef = TableReference.createWithEmptyNamespace(longTableName);
+        List<CharacterLimitType> kvsList = new ArrayList<CharacterLimitType>();
+        kvsList.add(CharacterLimitType.CASSANDRA);
+        kvsList.add(CharacterLimitType.POSTGRES);
+        assertThatThrownBy(() ->
+                schema.addTableDefinition(longTableName, getSimpleTableDefinition(tableRef)))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(getErrorMessage(longTableName, kvsList));
+    }
+
+    @Test
+    public void testLongTableNameLengthFailsNamespace() throws IOException {
+        // If namespace is non-empty, internal table name length is |namespace| + |tableName| + 2
+        Schema schema = new Schema("Table", TEST_PACKAGE, Namespace.DEFAULT_NAMESPACE);
+        String longTableName = String.join("", Collections.nCopies(40, "x"));
+        TableReference tableRef = TableReference.create(Namespace.DEFAULT_NAMESPACE, longTableName);
+        List<CharacterLimitType> kvsList = new ArrayList<CharacterLimitType>();
+        kvsList.add(CharacterLimitType.CASSANDRA);
+        assertThatThrownBy(() ->
+                schema.addTableDefinition(longTableName, getSimpleTableDefinition(tableRef)))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(getErrorMessage(longTableName, kvsList));
+    }
+
     private String readFileIntoString(File baseDir, String path) throws IOException {
         return new String(Files.toByteArray(new File(baseDir, path)), StandardCharsets.UTF_8);
     }
@@ -88,6 +145,14 @@ public class SchemaTest {
             columns();
             column("col1", "1", ValueType.VAR_LONG);
         }};
+    }
+
+    private String getErrorMessage(String tableName, List<CharacterLimitType> kvsExceeded) {
+        return String.format("Internal table name %s is too long, known to exceed character limits for " +
+                        "the following KVS: %s. If using a table prefix, please ensure that the concatenation " +
+                        "of the prefix with the internal table name is below the KVS limit. " +
+                        "If running only against a different KVS, set the ignoreTableNameLength flag.",
+                tableName, StringUtils.join(kvsExceeded, ", "));
     }
 
 }