+++ docs/source/release_notes/release-notes.rst
@@ -63,6 +63,10 @@ develop
            (`Pull Request <https://github.com/palantir/atlasdb/pull/2238>`__)
 
     *    - |devbreak|
+         - Throw an error during schema code generation stage if table length exceeds KVS limits.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/2220>`__)
+
+    *    - |devbreak|
          - IteratorUtils.forEach removed; it's not needed in a Java 8 codebase.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/2207>`__)